requirements.txt
# Core dependencies
aiohttp>=3.8.0
numpy>=1.21.0
sentence-transformers>=2.2.0

# Web framework
eel>=0.14.0

# Code editing and diff generation
pygments>=2.10.0  # For syntax highlighting

# Environment variables
python-dotenv>=0.19.0

# Optional: Development dependencies
black>=22.3.0  # Code formatting
pylint>=2.12.0  # Code linting
pytest>=7.0.0  # Testing
pytest-asyncio>=0.18.0  # Async testing support 